+++ genomepy/provider.py
@@ -26,16 +26,21 @@ from genomepy.utils import (
     read_url,
     safe,
     check_url,
+    retry,
     is_number,
     mkdir_p,
+    get_genomes_dir,
 )
 from genomepy.__about__ import __version__
 
-# store the output of slow commands (marked with @cached) for fast reuse
+# Store the output of slow commands (marked with @cache and @goldfish_cache) for fast reuse.
+# Bucketcache creates a new pickle for each function + set of unique variables,
+# but cannot ignore "self", so staticmethods are required.
 my_cache_dir = os.path.join(user_cache_dir("genomepy"), __version__)
 if not os.path.exists(my_cache_dir):
     os.makedirs(my_cache_dir)
-cached = Bucket(my_cache_dir, days=7)
+cache = Bucket(my_cache_dir, days=7)
+goldfish_cache = Bucket(my_cache_dir, minutes=10)
 
 config = norns.config("genomepy", default="cfg/default.yaml")
 
@@ -93,6 +98,13 @@ class ProviderBase(object):
             return subclass
 
         return decorator
+
+    @staticmethod
+    @goldfish_cache(ignore=["url", "max_tries"])
+    def provider_status(name, url, max_tries=1):
+        """check if provider is online (stores results for 10 minutes)"""
+        if not check_url(url, max_tries):
+            raise ConnectionError(f"{name} appears to be offline.\n")
 
     @classmethod
     def list_providers(cls):
@@ -173,7 +185,7 @@ class ProviderBase(object):
     def download_genome(
         self,
         name,
-        genomes_dir,
+        genomes_dir=None,
         localname=None,
         mask="soft",
         regex=None,
@@ -189,7 +201,7 @@ class ProviderBase(object):
         name : str
             Genome / species name
 
-        genomes_dir : str
+        genomes_dir : str , optional
             Directory to install genome
 
         localname : str , optional
@@ -208,20 +220,20 @@ class ProviderBase(object):
             If set to True the genome FASTA file will be compressed using bgzip.
             If not specified, the setting from the configuration file will be used.
         """
+        name = safe(name)
         self.check_name(name)
 
         link = self.get_genome_download_link(name, mask=mask, **kwargs)
 
-        original_name = name
-        name = safe(name)
         localname = get_localname(name, localname)
-
-        genomes_dir = os.path.expanduser(genomes_dir)
+        genomes_dir = get_genomes_dir(genomes_dir, check_exist=False)
         out_dir = os.path.join(genomes_dir, localname)
         if not os.path.exists(out_dir):
             mkdir_p(out_dir)
 
-        sys.stderr.write(f"Downloading genome from {link}...\n")
+        sys.stderr.write(
+            f"Downloading genome from {self.name}.\nTarget URL: {link}...\n"
+        )
 
         # download to tmp dir. Move genome on completion.
         # tmp dir is in genome_dir to prevent moving the genome between disks
@@ -290,7 +302,7 @@ class ProviderBase(object):
         metadata = {
             "name": localname,
             "provider": self.name,
-            "original name": original_name,
+            "original name": name,
             "original filename": os.path.split(link)[-1],
             "assembly_accession": self.assembly_accession(self.genomes.get(name)),
             "tax_id": self.genome_taxid(self.genomes.get(name)),
@@ -392,7 +404,9 @@ class ProviderBase(object):
             )
             return
 
-        sys.stderr.write(f"\nDownloading annotation from {link}...\n")
+        sys.stderr.write(
+            f"Downloading annotation from {self.name}.\nTarget URL: {link}...\n"
+        )
         try:
             self.download_and_generate_annotation(genomes_dir, link, localname)
         except Exception:
@@ -411,7 +425,7 @@ class ProviderBase(object):
         metadata["annotation url"] = link
         write_readme(readme, metadata, lines)
 
-    def download_annotation(self, name, genomes_dir, localname=None, **kwargs):
+    def download_annotation(self, name, genomes_dir=None, localname=None, **kwargs):
         """
         Download annotation file to to a specific directory
 
@@ -420,7 +434,7 @@ class ProviderBase(object):
         name : str
             Genome / species name
 
-        genomes_dir : str
+        genomes_dir : str , optional
             Directory to install annotation
 
         localname : str , optional
@@ -430,6 +444,8 @@ class ProviderBase(object):
 
         link = self.get_annotation_download_link(name, **kwargs)
 
+        localname = get_localname(name, localname)
+        genomes_dir = get_genomes_dir(genomes_dir, check_exist=False)
         self.attempt_and_report(name, localname, link, genomes_dir)
 
     def _search_taxids(self, genome, term):
@@ -443,6 +459,29 @@ class ProviderBase(object):
         for field in self.description_fields:
             if term in safe(genome[field].lower()):
                 return True
+
+    def _search_accessions(self, term):
+        """
+        Search for specific assembly accession.
+
+        Parameters
+        ----------
+        term : str
+            Assembly accession, GCA_....
+        Yields
+        ------
+        tuples with name and metadata
+        """
+        # NCBI provides a consistent assembly accession. This can be used to
+        # retrieve the species, and then search for that.
+        p = ProviderBase.create("NCBI")
+        species = [row[2] for row in p.search(term)]
+        if len(species) == 0:
+            raise ValueError(f"No genome found with accession {term}")
+        species = species[0]
+        for row in self.search(species):
+            if row[1] == term:
+                yield row
 
     def search(self, term):
         """
@@ -460,20 +499,26 @@ class ProviderBase(object):
         ------
         tuples with name and metadata
         """
-        term = str(term)
         genomes = self.genomes
-        if safe(term) in genomes:
-            yield self._genome_info_tuple(term)
+        term = safe(str(term))
+        if term.startswith("GCA_") and self.name != "NCBI":
+            for row in self._search_accessions(term):
+                yield (row)
 
         elif is_number(term):
             for name in genomes:
-                if self._search_taxids(genome=genomes[name], term=term):
+                if self._search_taxids(genomes[name], term):
                     yield self._genome_info_tuple(name)
 
+        elif term in genomes:
+            yield self._genome_info_tuple(term)
+
         else:
-            term = safe(term).lower()
+            term = term.lower()
             for name in genomes:
-                if self._search_descriptions(genome=genomes[name], term=term):
+                if term == safe(name).lower() or self._search_descriptions(
+                    genomes[name], term
+                ):
                     yield self._genome_info_tuple(name)
 
 
@@ -489,7 +534,7 @@ class EnsemblProvider(ProviderBase):
     The bacteria division is not yet supported.
     """
 
-    rest_url = "http://rest.ensembl.org/"
+    rest_url = "https://rest.ensembl.org/"
     provider_specific_install_options = {
         "toplevel": {
             "long": "toplevel",
@@ -505,11 +550,10 @@ class EnsemblProvider(ProviderBase):
     }
 
     def __init__(self):
-        # Necessary for bucketcache, otherwise methods with identical names
-        # from different classes will use the same cache :-O!
         self.name = "Ensembl"
+        self.provider_status(self.name, self.rest_url + "info/ping?", max_tries=2)
         # Populate on init, so that methods can be cached
-        self.genomes = self._get_genomes()
+        self.genomes = self._get_genomes(self._request_json, self.rest_url)
         self.accession_fields = ["assembly_accession"]
         self.taxid_fields = ["taxonomy_id"]
         self.description_fields = [
@@ -518,34 +562,32 @@ class EnsemblProvider(ProviderBase):
             "url_name",
             "display_name",
         ]
-        self.version = None
 
-    @cached(method=True)
-    def _request_json(self, ext):
+    @staticmethod
+    def _request_json(rest_url, ext):
         """Make a REST request and return as json."""
-        if self.rest_url.endswith("/") and ext.startswith("/"):
+        if rest_url.endswith("/") and ext.startswith("/"):
             ext = ext[1:]
 
-        r = requests.get(
-            self.rest_url + ext, headers={"Content-Type": "application/json"}
-        )
+        r = requests.get(rest_url + ext, headers={"Content-Type": "application/json"})
 
         if not r.ok:
             r.raise_for_status()
 
         return r.json()
 
-    @cached(method=True)
-    def _get_genomes(self):
+    @staticmethod
+    @cache
+    def _get_genomes(request_json, rest_url):
         sys.stderr.write("Downloading assembly summaries from Ensembl\n")
 
         genomes = {}
-        divisions = self._request_json("info/divisions?")
+        divisions = retry(request_json, 3, rest_url, "info/divisions?")
         for division in divisions:
             if division == "EnsemblBacteria":
                 continue
-            division_genomes = self._request_json(
-                "info/genomes/division/{}?".format(division)
+            division_genomes = retry(
+                request_json, 3, rest_url, f"info/genomes/division/{division}?"
             )
             for genome in division_genomes:
                 genomes[safe(genome["assembly_name"])] = genome
@@ -566,17 +608,14 @@ class EnsemblProvider(ProviderBase):
             genome.get("genebuild", "na"),
         )
 
-    def get_version(self, ftp_site):
-        """Retrieve current version from Ensembl FTP.
-        """
-        with urlopen(ftp_site + "/current_README") as response:
-            p = re.compile(r"Ensembl (Genomes|Release) (\d+)")
-            m = p.search(response.read().decode())
-        if m:
-            version = m.group(2)
-            sys.stderr.write("Using version {}\n".format(version))
-            self.version = version
-            return version
+    @staticmethod
+    @goldfish_cache(ignore=["request_json", "rest_url"])
+    def get_version(request_json, rest_url, vertebrates=False):
+        """Retrieve current version from Ensembl FTP."""
+        ext = "/info/data/?" if vertebrates else "/info/eg_version?"
+        ret = retry(request_json, 3, rest_url, ext)
+        releases = ret["releases"] if vertebrates else [ret["version"]]
+        return str(max(releases))
 
     def get_genome_download_link(self, name, mask="soft", **kwargs):
         """
@@ -604,14 +643,14 @@ class EnsemblProvider(ProviderBase):
 
         ftp_site = "ftp://ftp.ensemblgenomes.org/pub"
         if division == "vertebrates":
-            ftp_site = "http://ftp.ensembl.org/pub"
+            ftp_site = "ftp://ftp.ensembl.org/pub"
 
         # Ensembl release version
-        version = self.version
-        if kwargs.get("version"):
-            version = kwargs.get("version")
-        elif not version:
-            version = self.get_version(ftp_site)
+        version = kwargs.get("version")
+        if version is None:
+            version = self.get_version(
+                self._request_json, self.rest_url, division == "vertebrates"
+            )
 
         # division dependent url format
         ftp_dir = "{}/release-{}/fasta/{}/dna".format(
@@ -639,10 +678,10 @@ class EnsemblProvider(ProviderBase):
         # try to get the (much smaller) primary assembly,
         # unless specified otherwise
         link = get_url("primary_assembly")
-        if kwargs.get("toplevel") or not check_url(link):
+        if kwargs.get("toplevel") or not check_url(link, 2):
             link = get_url()
 
-        if check_url(link):
+        if check_url(link, 2):
             return link
 
         raise GenomeDownloadError(
@@ -670,14 +709,14 @@ class EnsemblProvider(ProviderBase):
 
         ftp_site = "ftp://ftp.ensemblgenomes.org/pub"
         if division == "vertebrates":
-            ftp_site = "http://ftp.ensembl.org/pub"
+            ftp_site = "ftp://ftp.ensembl.org/pub"
 
         # Ensembl release version
-        version = self.version
-        if kwargs.get("version"):
-            version = kwargs.get("version")
-        elif not version:
-            version = self.get_version(ftp_site)
+        version = kwargs.get("version")
+        if version is None:
+            version = self.get_version(
+                self._request_json, self.rest_url, division == "vertebrates"
+            )
 
         if division != "vertebrates":
             ftp_site += f"/{division}"
@@ -693,7 +732,7 @@ class EnsemblProvider(ProviderBase):
             version,
         )
 
-        if check_url(link):
+        if check_url(link, 2):
             return link
 
 
@@ -720,32 +759,32 @@ class UcscProvider(ProviderBase):
     }
 
     def __init__(self):
-        # Necessary for bucketcache, otherwise methods with identical names
-        # from different classes will use the same cache :-O!
         self.name = "UCSC"
+        self.provider_status(self.name, self.base_url)
         # Populate on init, so that methods can be cached
-        self.genomes = self._get_genomes()
+        self.genomes = self._get_genomes(self.rest_url)
         self.accession_fields = []
         self.taxid_fields = ["taxId"]
         self.description_fields = ["description", "scientificName"]
 
-    @cached(method=True)
-    def _get_genomes(self):
+    @staticmethod
+    @cache
+    def _get_genomes(rest_url):
         sys.stderr.write("Downloading assembly summaries from UCSC\n")
 
-        r = requests.get(self.rest_url, headers={"Content-Type": "application/json"})
+        r = requests.get(rest_url, headers={"Content-Type": "application/json"})
         if not r.ok:
             r.raise_for_status()
         ucsc_json = r.json()
         genomes = ucsc_json["ucscGenomes"]
         return genomes
 
-    # staticmethod, but the decorators cannot be combined
-    @cached(method=True)
-    def assembly_accession(self, genome):
+    @staticmethod
+    @cache
+    def assembly_accession(genome):
         """Return the assembly accession (GCA_*) for a genome.
 
-        UCSC does not server the assembly accession through the REST API.
+        UCSC does not serve the assembly accession through the REST API.
         Therefore, the readme.html is scanned for a GCA assembly id. If it is
         not found, the linked NCBI assembly page will be checked. Especially
         for older genome builds, the GCA will not be present, in which case
@@ -932,17 +971,23 @@ class NcbiProvider(ProviderBase):
     provider_specific_install_options = {}
 
     def __init__(self):
-        # Necessary for bucketcache, otherwise methods with identical names
-        # from different classes will use the same cache :-O!
         self.name = "NCBI"
+        self.provider_status(self.name, self.assembly_url)
         # Populate on init, so that methods can be cached
-        self.genomes = self._get_genomes()
+        self.genomes = self._get_genomes(self.assembly_url)
         self.accession_fields = ["assembly_accession", "gbrs_paired_asm"]
         self.taxid_fields = ["species_taxid", "taxid"]
-        self.description_fields = ["submitter", "organism_name"]
+        self.description_fields = [
+            "submitter",
+            "organism_name",
+            "assembly_accession",
+            "gbrs_paired_asm",
+            "paired_asm_comp",
+        ]
 
-    @cached(method=True)
-    def _get_genomes(self):
+    @staticmethod
+    @cache
+    def _get_genomes(assembly_url):
         """Parse genomes from assembly summary txt files."""
         sys.stderr.write(
             "Downloading assembly summaries from NCBI, this will take a while...\n"
@@ -957,7 +1002,7 @@ class NcbiProvider(ProviderBase):
         ]
         for fname in names:
             urlcleanup()
-            with urlopen(os.path.join(self.assembly_url, fname)) as response:
+            with urlopen(os.path.join(assembly_url, fname)) as response:
                 lines = response.read().decode("utf-8").splitlines()
             header = lines[1].strip("# ").split("\t")
             for line in lines[2:]:
@@ -1196,7 +1241,7 @@ class UrlProvider(ProviderBase):
         if check_url(link):
             return link
 
-    def download_annotation(self, url, genomes_dir, localname=None, **kwargs):
+    def download_annotation(self, url, genomes_dir=None, localname=None, **kwargs):
         """
         Attempts to download a gtf or gff3 file from the same location as the genome url
 
@@ -1219,6 +1264,7 @@ class UrlProvider(ProviderBase):
         """
         name = get_localname(url)
         localname = get_localname(name, localname)
+        genomes_dir = get_genomes_dir(genomes_dir, check_exist=False)
 
         if kwargs.get("to_annotation"):
             link = self.get_annotation_download_link(None, **kwargs)