+++ tests/06_provider_ensembl_test.py
@@ -4,7 +4,10 @@ import os
 import pytest
 import requests
 
+from platform import system
+
 travis = "TRAVIS" in os.environ and os.environ["TRAVIS"] == "true"
+linux = system() == "Linux"
 
 
 def validate_gzipped_gtf(fname):
@@ -35,7 +38,8 @@ def validate_gzipped_bed(fname):
 
 @pytest.fixture(scope="module")
 def p():
-    return genomepy.provider.EnsemblProvider()
+    p = genomepy.provider.EnsemblProvider()
+    return p
 
 
 def test_ensemblprovider__init__(p):
@@ -45,13 +49,13 @@ def test_ensemblprovider__init__(p):
 
 
 def test__request_json(p):
-    divisions = p._request_json("info/divisions?")
+    divisions = p._request_json("https://rest.ensembl.org/", "info/divisions?")
     assert isinstance(divisions, list)
     assert "EnsemblVertebrates" in divisions
 
     # test not r.ok
     with pytest.raises(requests.exceptions.HTTPError):
-        p._request_json("error")
+        p._request_json("https://rest.ensembl.org/", "error")
 
 
 def test__get_genomes(p):
@@ -71,32 +75,29 @@ def test_genome_info_tuple(p):
 
 
 def test_get_version(p):
-    # note: this test will break every time ensembl releases a new version
-    ftp_site = "http://ftp.ensembl.org/pub"
-    v = p.get_version(ftp_site)
-    assert v == "100"
+    # note: this test will break every time Ensembl releases a new version
+    v = p.get_version(p._request_json, "https://rest.ensembl.org/", True)
+    assert v == "101"
 
-    if not travis:
-        ftp_site = "ftp://ftp.ensemblgenomes.org/pub"
-        v = p.get_version(ftp_site)
-        assert v == "47"
+    v = p.get_version(p._request_json, "https://rest.ensembl.org/")
+    assert v == "48"
 
 
+@pytest.mark.skipif(travis and linux, reason="FTP does not work on Travis-Linux")
 def test_get_genome_download_link(p):
-    if not travis:
-        # non vertebrate: soft masked
-        link = p.get_genome_download_link("TAIR10", mask="soft", **{"version": 46})
-        assert (
-            link
-            == "ftp://ftp.ensemblgenomes.org/pub/plants/release-46/"
-            + "fasta/arabidopsis_thaliana/dna/Arabidopsis_thaliana.TAIR10.dna_sm.toplevel.fa.gz"
-        )
+    # non vertebrate: soft masked
+    link = p.get_genome_download_link("TAIR10", mask="soft", **{"version": 46})
+    assert (
+        link
+        == "ftp://ftp.ensemblgenomes.org/pub/plants/release-46/"
+        + "fasta/arabidopsis_thaliana/dna/Arabidopsis_thaliana.TAIR10.dna_sm.toplevel.fa.gz"
+    )
 
     # vertebrate with primary assembly: unmasked
     link = p.get_genome_download_link("GRCz11", mask="none", **{"version": 98})
     assert (
         link
-        == "http://ftp.ensembl.org/pub/release-98/fasta/"
+        == "ftp://ftp.ensembl.org/pub/release-98/fasta/"
         + "danio_rerio/dna/Danio_rerio.GRCz11.dna.primary_assembly.fa.gz"
     )
 
@@ -106,44 +107,45 @@ def test_get_genome_download_link(p):
     )
     assert (
         link
-        == "http://ftp.ensembl.org/pub/release-98/fasta/"
+        == "ftp://ftp.ensembl.org/pub/release-98/fasta/"
         + "danio_rerio/dna/Danio_rerio.GRCz11.dna_rm.toplevel.fa.gz"
     )
 
-    # vertebrate: any version
-    p.version = None
-    link = p.get_genome_download_link("GRCz11", **{"toplevel": True})
-    for substring in [
-        "http://ftp.ensembl.org/pub/release-",
-        "/fasta/danio_rerio/dna/Danio_rerio.GRCz11.dna_sm.toplevel.fa.gz",
-    ]:
-        assert substring in link
+    # vertebrate: latest version
+    version = p.get_version(p._request_json, "https://rest.ensembl.org/", True)
+    link = p.get_genome_download_link(
+        "GRCz11", **{"version": version, "toplevel": True}
+    )
+    expected_link = (
+        f"ftp://ftp.ensembl.org/pub/release-{version}/"
+        "fasta/danio_rerio/dna/Danio_rerio.GRCz11.dna_sm.toplevel.fa.gz"
+    )
+    assert link == expected_link
 
 
+@pytest.mark.skipif(travis and linux, reason="FTP does not work on Travis-Linux")
 def test_get_annotation_download_link(p):
-    if not travis:
-        # non vertebrate
-        link = p.get_annotation_download_link("TAIR10", **{"version": 46})
-        assert (
-            link
-            == "ftp://ftp.ensemblgenomes.org/pub/plants/release-46/"
-            + "gtf/arabidopsis_thaliana/Arabidopsis_thaliana.TAIR10.46.gtf.gz"
-        )
+    # non vertebrate
+    link = p.get_annotation_download_link("TAIR10", **{"version": 46})
+    expected_link = (
+        "ftp://ftp.ensemblgenomes.org/pub/plants/release-46/"
+        "gtf/arabidopsis_thaliana/Arabidopsis_thaliana.TAIR10.46.gtf.gz"
+    )
+    assert link == expected_link
 
     # vertebrate
     link = p.get_annotation_download_link("GRCz11", **{"version": 98})
-    assert (
-        link
-        == "http://ftp.ensembl.org/pub/release-98/gtf/"
-        + "danio_rerio/Danio_rerio.GRCz11.98.gtf.gz"
+    expected_link = (
+        "ftp://ftp.ensembl.org/pub/release-98/gtf/"
+        "danio_rerio/Danio_rerio.GRCz11.98.gtf.gz"
     )
+    assert link == expected_link
 
-    # vertebrate: any version
-    p.version = None
-    link = p.get_annotation_download_link("GRCz11")
-    for substring in [
-        "http://ftp.ensembl.org/pub/release-",
-        "/gtf/danio_rerio/Danio_rerio.GRCz11.",
-        ".gtf.gz",
-    ]:
-        assert substring in link
+    # vertebrate: latest version
+    version = p.get_version(p._request_json, "https://rest.ensembl.org/", True)
+    link = p.get_annotation_download_link("GRCz11", **{"version": version})
+    expected_link = (
+        f"ftp://ftp.ensembl.org/pub/release-{version}/"
+        f"gtf/danio_rerio/Danio_rerio.GRCz11.{version}.gtf.gz"
+    )
+    assert link == expected_link